requirements.txt
torch
torchvision
opencv-python
pyzmq
av
transformers
numpy
matplotlib
seaborn
scipy